[Kernels] Fix aligned_malloc conflict on Windows (#450)
* fix win aligned_malloc conflict

* avoid void pointer arithmetic

* avoid rand_r

* avoid false positive mem leak error of klocwork

* fix klocwork issues

* cpplint happy

* fix C3016: index variable in OpenMP 'for' statement must have signed integral type

* avoid dump jit_code to stdout


* make gcc happy in release build

* disable attention test on windows for now

+++ intel_extension_for_transformers/backends/neural_engine/kernels/src/kernels/spmm_avx512f.cpp
@@ -88,7 +88,7 @@ bool spmm_avx512f_k_t::init() {
 
 bool spmm_avx512f_k_t::execute(const std::vector<const void*>& rt_data) const {
 #pragma omp parallel for
-  for (size_t i = 0; i < jit_kers_.size(); ++i) {
+  for (int i = 0; i < static_cast<int>(jit_kers_.size()); ++i) {
     auto& jit_impl = jit_kers_[i];
     ssd::avx512_data_t rt_param;
     rt_param.sparse = jit_impl->bsc_data()->data().data();
